fix: homework3

See merge request amr_stacenko/minibank!3

+++ Minibank.Core/IDatabase.cs
@@ -2,6 +2,6 @@ namespace Minibank.Core
 {
     public interface IDatabase
     {
-        public int Get(string currencyCode);
+        public double Get(string currencyCode);
     }
 }
